requirements.txt
aiogram>=3.0.0
aiohttp
pandas
numpy
python-dotenv
